Implement basic Emscripten runner
em-interp is a small, wasmjit-omr-based WebAssembly runner (interpreter)
that is capable of executing WebAssembly generated by Emscripten.
Essentially, it provides a host module with the necessary imports to
support Emscripten output.

Currently, only a linear memory and the `_time` syscall are implemented.

Also, when attempting to import an unknown function, an message is
printed to the screen and a "dummy" that will trap when called is
registered as the callback for the import.